I have added logging of the business logic. Now all service methods are checked and if they cannot perform a task, that information is logged and problems are dealt with.

+++ src/main/java/com/adtimokhin/services/report/impl/ReportServiceImpl.java
@@ -9,6 +9,7 @@ import com.adtimokhin.services.comment.CommentService;
 import com.adtimokhin.services.report.ReportService;
 import com.adtimokhin.services.topic.TopicService;
 import com.adtimokhin.services.user.UserService;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -40,6 +41,9 @@ public class ReportServiceImpl implements ReportService {
     private TopicService topicService;
 
 
+    private static final Logger logger = Logger.getLogger("file");
+    private static final Logger adminLogger = Logger.getLogger("admin");
+
     @Override
     public List<Report> getAllReports() {
         return repository.findAll();
@@ -52,6 +56,10 @@ public class ReportServiceImpl implements ReportService {
 
     @Override
     public List<Report> getAllReportsByReportedUser(User user) {
+        if (user == null) {
+            logger.info("Tried to get reports by a null reported user");
+            return null;
+        }
         return repository.findAllByReportedUser(user);
     }
 
@@ -80,16 +88,38 @@ public class ReportServiceImpl implements ReportService {
 
     @Override
     public void banUser(Report report, String reason, User admin) {
-        if (admin.getRoles().contains(Role.ROLE_ADMIN)){
+        if (admin == null) {
+            logger.info("Null user tried to ban another user");
+            adminLogger.info("The report was malformed");
+            return;
+        }
+        if (report == null) {
+            logger.info("User with id " + admin.getId() + " tried to ban another user");
+            adminLogger.info("The report was malformed");
+            return;
+        }
+
+        if (admin.getRoles().contains(Role.ROLE_ADMIN)) {
             userService.banUser(report.getReportedUser());
             repository.delete(report);
+        } else {
+            logger.info("A user with id " + admin.getId() + " tried to block a user for a report with id " + report.getId() + " thought that user is not an ADMIN");
+            adminLogger.info("The report was malformed");
         }
     }
 
     @Override
     public void unBanUser(User user, User admin) {
-        if(admin.getRoles().contains(Role.ROLE_ADMIN)){
+        if (user == null || admin == null) {
+            logger.info("Either admin or a unbanned user was a null.");
+            adminLogger.info("The report was malformed");
+            return;
+        }
+
+        if (admin.getRoles().contains(Role.ROLE_ADMIN)) {
             userService.unBanUser(user);
+        } else {
+            logger.info("A user with id " + admin.getId() + " tried to unban user with id " + user.getId() + ", though that user was not an ADMIN");
         }
     }
 }
